backend/scraping/requirements.txt
pika==1.2.0
pymongo==4.0.1
python-dotenv==0.19.2
